backend/device-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1
redis==5.0.1
pydantic==2.5.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
python-keycloak==3.7.0
httpx==0.25.2
pydantic-settings==2.1.0
paho-mqtt==1.6.1
asyncio-mqtt==0.16.2
websockets==12.0